Make edit mode section into more generic part because it will be use in the other part.
